+++ src/Phan/Config/Initializer.php
@@ -285,6 +285,7 @@ EOT;
                 'SleepCheckerPlugin',
                 'UnreachableCodePlugin',
                 'UseReturnValuePlugin',
+                'EmptyStatementListPlugin',
             ];
         }